cluster: extract config replication path into sync_conn_config.go
Fourth pure code-motion increment of the #5661 sync_conn.go split. Moves
the HA config generation queue and the apply loop into a cohesive sibling
file.

Moved verbatim (no rename, no logic/timing change):
  nextConfigGen, QueueConfig, shouldApplyConfigGen, recordAppliedConfigGen,
  configApplyLoop.

Same package (pkg/cluster); imports recomputed per file. Behavior is
byte-identical.

go build ./... passes.
