+++ pkg/cluster/sync_conn.go
@@ -385,131 +385,6 @@ func (s *SessionSync) Stop() {
 	}
 }
 
-// nextConfigGen draws the next strictly-monotonic config generation stamped
-// on an outgoing config-sync message (#3931). The counter is seeded from
-// CLOCK_MONOTONIC nanos at construction so it never regresses below a value
-// the peer may already hold across this node's restarts within a boot.
-func (s *SessionSync) nextConfigGen() uint64 {
-	return s.configGenCounter.Add(1)
-}
-
-// QueueConfig sends the full config text to the peer for configuration
-// synchronization. The payload carries a monotonic config generation (#3931)
-// so the receiver can order a rapid commit pair and refuse a reordered older
-// config.
-func (s *SessionSync) QueueConfig(configText string) {
-	conn := s.getActiveConn()
-	if conn == nil {
-		return
-	}
-	gen := s.nextConfigGen()
-	payload := encodeConfigPayload(configText, gen)
-	s.writeMu.Lock()
-	err := writeMsg(conn, syncMsgConfig, payload)
-	s.writeMu.Unlock()
-	if err != nil {
-		slog.Warn("cluster sync: config send error", "err", err)
-		s.stats.Errors.Add(1)
-		s.handleDisconnect(conn)
-		return
-	}
-	s.stats.ConfigsSent.Add(1)
-	slog.Info("cluster sync: config sent to peer", "size", len(configText), "gen", gen)
-}
-
-// shouldApplyConfigGen is the admission half of the #3931 receiver-side config
-// ordering guard. It returns true if a config with generation gen should be
-// ATTEMPTED — gen==0 (a legacy sender / unknown, applied unconditionally) or
-// strictly newer than the last SUCCESSFULLY-applied generation — and false if
-// it is an out-of-order older / duplicate config that must be dropped.
-//
-// It deliberately does NOT advance the high-water mark; the caller advances via
-// recordAppliedConfigGen ONLY after the apply succeeds (M-2/#4151). Advancing
-// on admission (the pre-#4151 behavior) meant an apply failure left the
-// high-water ahead of the actually-applied config, so the primary's re-push of
-// the same generation was dropped as stale and the standby stayed silently
-// stranded on the prior config. It is called ONLY from the single-consumer
-// configApplyLoop, so the load needs no CAS.
-func (s *SessionSync) shouldApplyConfigGen(gen uint64) bool {
-	if gen == 0 {
-		return true
-	}
-	last := s.lastAppliedConfigGen.Load()
-	return last == 0 || gen > last
-}
-
-// recordAppliedConfigGen advances the config high-water mark after a config of
-// generation gen has been SUCCESSFULLY applied. gen==0 (legacy / unconditional)
-// never advances the mark, mirroring the pre-#3931 behavior. It is called ONLY
-// from the single-consumer configApplyLoop after OnConfigReceived returns nil,
-// so the load/store pair needs no CAS.
-func (s *SessionSync) recordAppliedConfigGen(gen uint64) {
-	if gen == 0 {
-		return
-	}
-	if gen > s.lastAppliedConfigGen.Load() {
-		s.lastAppliedConfigGen.Store(gen)
-	}
-}
-
-// configApplyLoop is the single ordered consumer of config-sync messages
-// (#3931). It drains configApplyCh in receive order and attempts a config only
-// when shouldApplyConfigGen accepts its generation, so a reordered older config
-// from a rapid commit pair (C1 after C2) is dropped and the standby always
-// converges to the newest config. Replaces the pre-#3931 racing
-// `go OnConfigReceived` per message.
-//
-// The config high-water mark (lastAppliedConfigGen) advances ONLY after the
-// apply succeeds (recordAppliedConfigGen, gated on OnConfigReceived returning
-// nil). An apply failure counts ConfigsApplyFailed and leaves the high-water at
-// the last-applied generation, so the primary's re-push of the same generation
-// is re-admitted and the standby re-converges instead of being silently
-// stranded on the prior config (M-2/#4151).
-func (s *SessionSync) configApplyLoop(ctx context.Context) {
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case item := <-s.configApplyCh:
-			if !s.shouldApplyConfigGen(item.gen) {
-				s.stats.ConfigsStaleIgnored.Add(1)
-				slog.Warn("cluster sync: dropping out-of-order config sync (stale generation) — standby retains newer config",
-					"incoming_gen", item.gen, "last_applied_gen", s.lastAppliedConfigGen.Load(), "size", len(item.text))
-				continue
-			}
-			if s.OnConfigReceived == nil {
-				// No apply handler wired — the config cannot be applied, so the
-				// high-water must NOT advance (M-2/#4151). A later wired handler
-				// re-applies on the primary's next push of this generation.
-				continue
-			}
-			if err := s.OnConfigReceived(item.text); err != nil {
-				// The apply did not take effect (compile/promote failure or a
-				// transient RG0-primary rejection). Do NOT advance the
-				// high-water: leaving it at the last-applied generation keeps
-				// the standby eligible for the primary's re-push so it
-				// re-converges instead of being silently stranded (M-2/#4151).
-				s.stats.ConfigsApplyFailed.Add(1)
-				// Debug, not Warn: handleConfigSync (the callback) already logs
-				// the authoritative one-time reason WHY the apply failed
-				// (RG0-primary rejection at Warn, compile/promote failure at
-				// Error). This line is the per-retry diagnostic detail of the
-				// M-2/#4151 high-water retention, and the same generation may be
-				// re-pushed on every reconnect while the condition persists — a
-				// Warn here would spam (CLAUDE.md logging rule). The
-				// rate-independent observable is the ConfigsApplyFailed counter
-				// above (surfaced in cluster status), not this log.
-				slog.Debug("cluster sync: config apply failed — retaining prior high-water so the peer re-push re-converges",
-					"incoming_gen", item.gen, "last_applied_gen", s.lastAppliedConfigGen.Load(), "size", len(item.text), "err", err)
-				continue
-			}
-			// Apply confirmed — advance the high-water so a duplicate re-push of
-			// the same generation is correctly skipped as stale.
-			s.recordAppliedConfigGen(item.gen)
-		}
-	}
-}
-
 func (s *SessionSync) acceptLoop(ctx context.Context, ln net.Listener, fabricIdx int) {
 	for {
 		conn, err := ln.Accept()
